feat: implementar catálogo de OVAs con formulario de creación

+++ ova-frontend/src/app/components/ova-catalog/ova-catalog.ts
@@ -1,13 +1,24 @@
 import { Component, OnInit } from '@angular/core';
-import { OvaService } from '../../services/ova.service';
+import { OvaService, Ova } from '../../services/ova.service';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   selector: 'app-ova-catalog',
+  standalone: true,
+  imports: [CommonModule, FormsModule],
   templateUrl: './ova-catalog.html',
   styleUrls: ['./ova-catalog.scss']
 })
 export class OvaCatalogComponent implements OnInit {
-  ovas: any[] = [];
+  ovas: Ova[] = [];
+  showForm = false;
+  newOva: Ova = {
+    titulo: '',
+    descripcion: '',
+    autor: '',
+    fechaCreacion: ''
+  };
 
   constructor(private ovaService: OvaService) {}
 
@@ -21,9 +32,37 @@ export class OvaCatalogComponent implements OnInit {
     });
   }
 
-  deleteOva(id: number) {
-    this.ovaService.deleteOva(id).subscribe(() => {
-      this.loadOvas();
-    });
+  toggleForm() {
+    this.showForm = !this.showForm;
+    if (!this.showForm) {
+      this.resetForm();
+    }
+  }
+
+  createOva() {
+    if (this.newOva.titulo && this.newOva.descripcion && this.newOva.autor) {
+      this.newOva.fechaCreacion = new Date().toISOString().split('T')[0];
+      this.ovaService.createOva(this.newOva).subscribe(() => {
+        this.loadOvas();
+        this.toggleForm();
+      });
+    }
+  }
+
+  deleteOva(id: number | undefined) {
+    if (id && confirm('¿Está seguro de eliminar este OVA?')) {
+      this.ovaService.deleteOva(id).subscribe(() => {
+        this.loadOvas();
+      });
+    }
+  }
+
+  resetForm() {
+    this.newOva = {
+      titulo: '',
+      descripcion: '',
+      autor: '',
+      fechaCreacion: ''
+    };
   }
 }
